+++ frontend/src/app/[locale]/(main)/dashboard/components/DashboardStats.tsx
@@ -1,6 +1,6 @@
 import { useTranslations } from 'next-intl';
 import StatCard from '@/components/common/StatCard';
-import { UserStatsResponse } from '@/interfaces/UserStats';
+import { UserStatsResponse } from '@/interfaces/models/UserStats';
 import { use } from 'react';
 
 export default function DashboardStats({ stats }: UserStatsResponse) {
