+++ frontend/src/app/[locale]/(main)/feedback/[id]/components/FeedbackDetail.tsx
@@ -17,7 +17,7 @@ import {
   AccordionItem,
   AccordionTrigger,
 } from '@/components/ui/Accordion';
-import { FeedbackResponse } from '@/interfaces/SessionFeedback';
+import { FeedbackResponse } from '@/interfaces/models/SessionFeedback';
 import { useCallback, useEffect, useState } from 'react';
 import { useTranslations } from 'next-intl';
 import { getSessionFeedback } from '@/services/client/SessionService';
