+++ frontend/src/app/[locale]/(main)/history/components/HistoryStats.tsx
@@ -1,7 +1,7 @@
 import { useTranslations } from 'next-intl';
 import Progress from '@/components/ui/Progress';
 import StatCard from '@/components/common/StatCard';
-import { UserStatsResponse } from '@/interfaces/UserStats';
+import { UserStatsResponse } from '@/interfaces/models/UserStats';
 import { use } from 'react';
 
 export default function HistoryStats({ stats }: UserStatsResponse) {
