+++ frontend/src/app/[locale]/(main)/settings/components/Settings.tsx
@@ -14,9 +14,9 @@ import Switch from '@/components/ui/Switch';
 import UserConfidenceFields from '@/components/common/UserConfidenceFields';
 import { DeleteUserHandler } from '@/components/common/DeleteUserHandler';
 import { UserProfileService } from '@/services/client/UserProfileService';
-import { UserPreference } from '@/interfaces/inputs/UserInputFields';
+import { UserPreference } from '@/interfaces/models/UserInputFields';
 import { PrimaryGoals, UserRoles } from '@/lib/utils';
-import { UserProfile } from '@/interfaces/UserProfile';
+import { UserProfile } from '@/interfaces/models/UserProfile';
 import { showErrorToast, showSuccessToast } from '@/lib/toast';
 import JSZip from 'jszip';
 import UserPreferences from './UserPreferences';
