--- frontend/src/interfaces/inputs/Form.ts
@@ -1,12 +1,0 @@
-import { type FieldPath, type FieldValues } from 'react-hook-form';
-
-export type FormFieldContextValue<
-  TFieldValues extends FieldValues = FieldValues,
-  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
-> = {
-  name: TName;
-};
-
-export type FormItemContextValue = {
-  id: string;
-};
