--- frontend/src/interfaces/inputs/PasswordInput.ts
@@ -1,14 +1,0 @@
-import { InputHTMLAttributes } from 'react';
-
-export interface PasswordRequirement {
-  id: string;
-  label: string;
-  test: (password: string) => boolean;
-}
-
-export interface PasswordInputProps extends Omit<InputHTMLAttributes<HTMLInputElement>, 'type'> {
-  placeholder: string;
-  disabled: boolean;
-  requirements?: PasswordRequirement[];
-  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-}
