--- frontend/src/interfaces/inputs/PasteField.ts
@@ -1,4 +1,0 @@
-export interface PasteField {
-  onChange: (code: string) => void;
-  value?: string;
-}
