--- frontend/src/interfaces/inputs/Rating.ts
@@ -1,34 +1,0 @@
-import type { KeyboardEvent, MouseEvent, ReactElement, ReactNode } from 'react';
-import { type LucideProps } from 'lucide-react';
-
-export type RatingContextValue = {
-  value: number;
-  readOnly: boolean;
-  hoverValue: number | null;
-  focusedStar: number | null;
-  handleValueChange: (
-    event: MouseEvent<HTMLButtonElement> | KeyboardEvent<HTMLButtonElement>,
-    value: number
-  ) => void;
-  handleKeyDown: (event: KeyboardEvent<HTMLButtonElement>) => void;
-  setHoverValue: (value: number | null) => void;
-  setFocusedStar: (value: number | null) => void;
-};
-
-export type RatingButtonProps = LucideProps & {
-  index?: number;
-  icon?: ReactElement<LucideProps>;
-};
-
-export type RatingProps = {
-  defaultValue: number;
-  value?: number;
-  onChange?: (
-    event: MouseEvent<HTMLButtonElement> | KeyboardEvent<HTMLButtonElement>,
-    value: number
-  ) => void;
-  onValueChange?: (value: number) => void;
-  readOnly?: boolean;
-  className?: string;
-  children?: ReactNode;
-};
