+++ frontend/src/interfaces/models/ConversationScenario.ts
