+++ frontend/src/interfaces/models/PrivacyDialog.ts
