+++ frontend/src/interfaces/models/UserStats.ts
