+++ frontend/src/interfaces/store/ConversationScenarioState.ts
